# contracts/api/openapi.yaml
openapi: 3.1.0
info:
  title: TUBA Product API
  version: 0.1.0
  description: Current API contract. Operations marked planned are design contracts and are not implemented endpoints.
servers:
  - url: /api/v1
security:
  - oidc: []
paths:
...
  /collectors:
    get:
      operationId: listCollectors
      x-required-permission: source:manage
      summary: List registered Collectors in the authenticated tenant
      parameters:
        - {$ref: "#/components/parameters/Cursor"}
        - {$ref: "#/components/parameters/Limit"}
      responses:
        "200":
          description: Collector registry and last reported status
          content: {application/json: {schema: {type: object, required: [items], properties: {items: {type: array, items: {$ref: "#/components/schemas/CollectorSummary"}}}}}}
        "403": {$ref: "#/components/responses/Forbidden"}
        "503": {description: Collector registry unavailable}
  /collectors/enrollments:
    post:
      operationId: createCollectorEnrollment
      x-required-permission: source:manage
      summary: Create a one-time enrollment token for a Collector
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              properties:
                expires_in_minutes: {type: integer, minimum: 5, maximum: 1440, default: 30}
      responses:
        "201":
          description: Token returned once; response is no-store
          content: {application/json: {schema: {$ref: "#/components/schemas/CollectorEnrollment"}}}
        "400": {description: Invalid expiration}
        "403": {$ref: "#/components/responses/Forbidden"}
        "503": {description: Collector registry unavailable}
...
  /collectors/{id}:
    delete:
      operationId: disableCollector
      x-required-permission: source:manage
      summary: Disable a Collector credential
      parameters:
        - {name: id, in: path, required: true, schema: {type: string, pattern: '^col_[a-f0-9]{32}$'}}
      responses:
        "204": {description: Collector disabled}
        "403": {$ref: "#/components/responses/Forbidden"}
        "404": {description: Collector not found or already disabled}
        "503": {description: Collector registry unavailable}
  /collector/enroll:
    post:
      operationId: enrollCollector
      summary: Consume a one-time enrollment token and register this Collector
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              required: [enrollment_token, install_id, hostname, os, architecture, version]
              properties:
                enrollment_token: {type: string, maxLength: 128}
                install_id: {type: string, minLength: 1, maxLength: 128}
                hostname: {type: string, minLength: 1, maxLength: 255}
                os: {type: string, enum: [linux, windows]}
                architecture: {type: string, enum: [amd64, arm64]}
                version: {type: string, minLength: 1, maxLength: 64}
      responses:
        "201": {description: Collector ID and bearer credential returned once; response is no-store}
        "400": {description: Invalid registration}
        "401": {description: Invalid, expired, or consumed enrollment token}
        "503": {description: Collector registry unavailable}
  /collector/heartbeat:
    post:
      operationId: reportCollectorHeartbeat
      summary: Report Collector health and bounded source counters
      security:
        - collectorBearer: []
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: "#/components/schemas/CollectorHeartbeat"}
      responses:
        "204": {description: Heartbeat accepted}
        "400": {description: Invalid heartbeat}
        "401": {$ref: "#/components/responses/Unauthorized"}
        "503": {description: Collector registry unavailable}
  /collector/config:
    get:
      operationId: getCollectorConfig
      summary: Poll the latest immutable configuration using ETag
      security:
        - collectorBearer: []
      parameters:
        - {name: If-None-Match, in: header, required: false, schema: {type: string}}
      responses:
        "200": {description: Latest configuration version and ETag, no-store}
        "204": {description: No configuration has been published}
        "304": {description: Configuration is unchanged}
        "401": {$ref: "#/components/responses/Unauthorized"}
        "503": {description: Collector registry unavailable}
...
  /analysis/feedback:
    post:
      operationId: createAnalysisFeedback
      x-required-permission: analysis:feedback
      summary: Record true-positive, false-positive, false-negative or inconclusive analyst feedback
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: "#/components/schemas/AnalysisFeedbackInput"}
      responses:
        "201":
          description: Feedback recorded
          content: {application/json: {schema: {$ref: "#/components/schemas/AnalysisFeedback"}}}
        "400": {description: Invalid feedback}
        "403": {$ref: "#/components/responses/Forbidden"}
components:
  securitySchemes:
    sourceApiKey:
      type: apiKey
      in: header
      name: X-API-Key
    collectorBearer:
      type: http
      scheme: bearer
      bearerFormat: TUBA Collector credential
    oidc:
      type: openIdConnect
      openIdConnectUrl: https://identity.example/realms/tuba/.well-known/openid-configuration
  parameters:
    Cursor:
      name: cursor
      in: query
      schema: {type: string, maxLength: 2048}
    Limit:
      name: limit
      in: query
      schema: {type: integer, minimum: 1, maximum: 200, default: 50}
  responses:
    Unauthorized:
      description: Authentication required or invalid
      content: {application/json: {schema: {$ref: "#/components/schemas/Error"}}}
    Forbidden:
      description: Effective permissions do not allow the operation
      content: {application/json: {schema: {$ref: "#/components/schemas/Error"}}}
  schemas:
...
    CollectorEnrollment:
      type: object
      required: [enrollment_token, expires_at]
      properties:
        enrollment_token: {type: string, description: One-time value; store securely and never log it}
        expires_at: {type: string, format: date-time}
    CollectorSummary:
      type: object
      required: [id, organization_id, namespace, hostname, os, architecture, installed_version, state, config_version, desired_config_version, online]
      properties:
        id: {type: string, pattern: '^col_[a-f0-9]{32}$'}
        organization_id: {type: string}
        namespace: {type: string}
        hostname: {type: string}
        os: {type: string}
        architecture: {type: string}
        installed_version: {type: string}
        desired_version: {type: string}
        state: {type: string}
        config_version: {type: integer}
        desired_config_version: {type: integer}
        last_heartbeat_at: {type: [string, "null"], format: date-time}
        online: {type: boolean}
    CollectorHeartbeat:
      type: object
      additionalProperties: false
      required: [version, config_version, state, sources, queue_depth]
      properties:
        version: {type: string, maxLength: 64}
        config_version: {type: integer, minimum: 0}
        state: {type: string, enum: [running, buffering, backpressured, paused, error]}
        queue_depth: {type: integer, minimum: 0}
        oldest_queued_at: {type: [string, "null"], format: date-time}
        diagnostic: {type: string, maxLength: 2048}
        sources:
          type: array
          maxItems: 256
          items:
            type: object
            additionalProperties: false
            required: [source_id, state, events_read, events_sent, events_drop]
            properties:
              source_id: {type: string, minLength: 1, maxLength: 128}
              state: {type: string, enum: [running, paused, error]}
              events_read: {type: integer, minimum: 0}
              events_sent: {type: integer, minimum: 0}
              events_drop: {type: integer, minimum: 0}
              last_error: {type: string, maxLength: 512}
...
    AnalysisFeedbackInput:
      type: object
      additionalProperties: false
      required: [feedback_type, reason]
      properties:
        feedback_type: {type: string, enum: [true_positive, false_positive, false_negative, inconclusive]}
        anomaly_id: {type: string, maxLength: 256}
        rule_id: {type: string, maxLength: 256}
        entity_id: {type: string, maxLength: 256}
        event_ids: {type: array, maxItems: 100, items: {type: string, maxLength: 256}}
        reason: {type: string, maxLength: 2000}
    AnalysisFeedback:
      type: object
      required: [id, feedback_type, event_ids, reason, created_at]
      properties:
        id: {type: integer}
        feedback_type: {type: string}
        anomaly_id: {type: string}
        rule_id: {type: string}
        entity_id: {type: string}
        event_ids: {type: array, items: {type: string}}
        reason: {type: string}
        created_at: {type: string, format: date-time}
...
    Error:
      type: object
      required: [code, message, request_id]
      properties:
        code: {type: string}
        message: {type: string}
        request_id: {type: string}